refactor:UnViewedCafeRepository 분리(#187)
