+++ server/src/main/java/com/project/yozmcafe/service/UnViewedCafeService.java
@@ -3,6 +3,7 @@ package com.project.yozmcafe.service;
 import com.project.yozmcafe.domain.cafe.Cafe;
 import com.project.yozmcafe.domain.cafe.CafeRepository;
 import com.project.yozmcafe.domain.cafe.UnViewedCafe;
+import com.project.yozmcafe.domain.cafe.UnViewedCafeRepository;
 import com.project.yozmcafe.domain.member.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,15 +15,17 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class UnViewedCafeService {
 
+    private final UnViewedCafeRepository unViewedCafeRepository;
     private final CafeRepository cafeRepository;
 
-    public UnViewedCafeService(final CafeRepository cafeRepository) {
+    public UnViewedCafeService(final UnViewedCafeRepository unViewedCafeRepository, final CafeRepository cafeRepository) {
+        this.unViewedCafeRepository = unViewedCafeRepository;
         this.cafeRepository = cafeRepository;
     }
 
     @Transactional
     public void removeUnViewedCafe(final Member member, final long cafeId) {
-        final UnViewedCafe unViewedCafe = cafeRepository.findUnViewedCafeByMemberAndCafe(member.getId(), cafeId)
+        final UnViewedCafe unViewedCafe = unViewedCafeRepository.findByMemberIdAndCafeId(member.getId(), cafeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당하는 UnViewedCafe가 없습니다."));
         member.removeUnViewedCafe(unViewedCafe);
