--- server/src/test/java/com/project/yozmcafe/util/UnViewedCafeRepository.java
@@ -1,7 +1,0 @@
-package com.project.yozmcafe.util;
-
-import com.project.yozmcafe.domain.cafe.UnViewedCafe;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UnViewedCafeRepository extends JpaRepository<UnViewedCafe, Long> {
-}
